+++ app/models/Strapi.ts
@@ -1,16 +1,21 @@
-import { StrapiApi } from 'hooks/StrapiAPI';
+import { StrapiApi } from "hooks/StrapiAPI";
 
 /**
   Model interfaces are a subtype of Strapi api responses that only include properties needed by consuming components
 */
 export namespace StrapiModel {
-  export interface Link extends StrapiApi.LinkResponse { }
-  export interface DynamicLink extends Omit<StrapiApi.DynamicLinkResponse, "__component"> { }
-  export interface Footer extends StrapiApi.FooterResponse { }
-  export interface Header extends StrapiApi.HeaderResponse { }
-  export interface Logo extends Pick<StrapiApi.LogoResponse, "width" | "height" | "alternativeText" | "url"> { }
-  export interface NavigationMenu extends StrapiApi.NavigationMenuResponse {
-  }
+  export interface Link extends StrapiApi.LinkResponse {}
+  export interface DynamicLink
+    extends Omit<StrapiApi.DynamicLinkResponse, "__component"> {}
+  export interface Footer extends StrapiApi.FooterResponse {}
+  export interface Header extends StrapiApi.HeaderResponse {}
+  export interface Logo
+    extends Pick<
+      StrapiApi.LogoResponse,
+      "width" | "height" | "alternativeText" | "url"
+    > {}
+  export interface NavigationMenu
+    extends Omit<StrapiApi.NavigationMenuResponse, "__component"> {}
 }
 
 export function extractNavigationMenusFromNavigationResponse(
